Move FeatureLoader to package ol.featureloader
