+++ gwt-ol3-client/src/main/java/ol/source/VectorOptions.java
@@ -21,6 +21,7 @@ import jsinterop.annotations.JsType;
 
 import ol.Collection;
 import ol.Feature;
+import ol.featureloader.FeatureLoader;
 
 /**
  * Vector source options.
